requirements/test.txt
-r prod.txt

pytest
tox
coverage
pytest-cov
pyhamcrest